Factor out blutspenden.de tools
